Chose a LICENSE
CC BY-SA stands most firmly behind what we feel makes sense for the work that we put out into the world. We want hobbyists to benefit from making knowledge public and simple, but also want to hold the rest of the CNC industry accountable to keeping that information public if Sienci were to be passed up. This, we feel, provides the most guarantee of long-term benefit to an industry that we're passionate about :)